Refactor model-qa (#253)
