+++ .gitignore
@@ -23,3 +23,5 @@ examples/python/hf_cache
 
 !test/test_models/hf-internal-testing/
 !test/test_models/hf-internal-testing/tiny-random-gpt2*/*.onnx
+
+.ipynb_checkpoints/
